Make h5py imports fail late.
This may sound crazy but there is a rationale behind it: in the current
neurom.io design, SWC and HDF5 readers are imported early. But HDF5 may
be considered an internal BBP format, and most external users may not
need it. Ideally, the IO should use a plug-in architecture to make the
optionality of readers more explicit.

TODO: This isn't correctly tested in the absence of h5py. Figure out how
to test it.