Simplify section iteration and mapping.
Add a function decorator section_function that holds the iteration
type. This pattern should allow to have a single iteration entry
point. For example,

    @section_type
    def foo(sec):
        pass

    @segment_type
    def bar(seg):
        pass

    nrn = ...
    sec_foos = [f for f in neurom.magic_iter(nrn, foo)]
    seg_bars = [b for b in neurom.magic_iter(nrn, bar)]

TODO: currently, the decorator to_val must be used in addition to
section_function:

    @section_type
    @to_val
    def baz(sec):
        pass

A possible improvement is to remove the need for to_val and add a
decorator parameter for functions that require tree objects:

    @section_type(tree_arg=True)
    def path_length(sec):
      return 42