Add example of section tree from SWC data block.
This example builds a section tree based on the SWC-like
internal data representation, meaning it can be used for any
of the input data formats. Given that there si a more efficient
reader for h5 data, this could initially be used for SWC and
NeuroLucida (NL) .asc files. Note that NL already comes in a
section-based tree structure, so a more efficient implementation
is also possible.