New section tree node type.
Class Section, is a tree node with a points attribute, holding
a 2D numpy array of points. This combines the previously disjoint
concepts of "section" as a collection of points, and tree node as
a generic element of a tree. Functions written to operate on sections
can now access the points and the tree hierarchy.

Resolves #421.