fix: ensure SWC reader makes soma sections.
Soma sections were not being detected when soma had only one emerging
branch. This is because the method to detect sections is topological.
New algorithm treats soma specially, creating a section when there is
a change of type from soma to something else, regardless of whether
a bifurcation point arises.

Fixes #463