Use decorator to register feature
* New feature:

The features are now registered as a hard coded list of functions but as follows:

```python
@feature(shape=[Shape.OnePerSegment], namespace='NEURITEFEATURES')
def segment_meander_angles(neurites, neurite_type=NeuriteType.all):
    """Inter-segment opening angles in a section."""
    ...
```

The 'shape' argument of the decorator holds the shape (a la numpy) of the
output result of the function.

* Deprecation

`register_neurite_feature` is now deprecated and is replaced by the `@feature` decorator.

Note:
This commit will be used in a following PR to fix:
https://github.com/BlueBrain/NeuroM/issues/859