💥 Change the shape of the output DataFrame of `extract_dataframe`
`extract_dataframe` used to return `N` rows for each neuron where `N` is the
number of neurite types in the input config.
per neuron.

Since `extract_dataframe` input is a list of neuron, it is more natural to
return exactly one row per neuron.

Use a MultiIndex on the columns to hold the metrics for each neurite type
instead of having the neurite type appear in the index (axis=0).

Before:

|   | name    | neurite_type    | metric A | metric B |
|---|---------|-----------------|----------|----------|
| 0 | no-axon | axon            | val A    | val B    |
| 1 | no-axon | apical_dendrite | val A'   | val B'   |

After:

|   |         | axon                | apical_dendrite            |
|   | name    | metric A | metric B | metric A        | metric B |
|---|---------|----------|----------|-----------------|----------|
| 0 | no-axon | val A    | val B    | val A'          | val B'   |