Fix various morph_stats issues
* Fixed issue #859: features in `morph_stats` that are None now have the correct shape

* Fixed issue with computing the mean, std and median of empty features

When features are returning an empty array, the mean, median and std of those
values will be set to None instead of crashing.

* Fixed issue with the convex hull computation in feature `neurite_volume_density`

The neurite volume computation made using a convex hull is now wrapped in a
try/catch block and the feature will return `np.nan` if the convex hull crashes.

* <cleanup> Removed no longer relevant NOTE comment dating from when we had to
maintain the compatibility between neurom.features and neurom.fst