Make deprecation warnings respect pre-existing warnings configuration (#1002)
The NeuroM deprecation warning configuration was hijacking the warnings setup in
order to enable visibility of the DeprecationWarning. This change doesn't enforce a 
warning filter configuration, but increases the stack level of the warn function to 3.

Setting it to 3 brings the warning stacl level to the __main__ (e.g. user's interactive
 shell) where the visibility of the DeprecationWarning is by default enabled.